Rename linear_transformation module
